requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==3.0.6
django-formtools==2.2
django-oauth-toolkit==1.3.2
django-otp==0.9.1
django-otp-yubikey==0.5.2
django-phonenumber-field==3.0.1
django-timedeltafield==0.7.10
django-two-factor-auth==1.11.0
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
phonenumbers==8.12.4
pycryptodome==3.9.7
pytz==2020.1
qrcode==6.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
YubiOTP==0.2.2.post1
